cherry-picked changes. test=release/1.3
